Add commands

+++ Models/Consumer.cs
@@ -8,11 +8,11 @@ namespace WaterNetworkProject.Models
 {
     public class Consumer
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
 
-        public Consumer(int id, string firstName, string lastName)
+        public Consumer(long id, string firstName, string lastName)
         {
             Id = id;
             FirstName = firstName;
